Refactor to use PascalCase for component filenames
Renamed component files and updated imports throughout the codebase to use PascalCase naming convention for consistency and improved readability. This affects UI, sidebar, splash screen, wallet, and other component directories. No functional changes were made.
