+++ app/components/DeleteConfirmationDialog.tsx
@@ -1,24 +1,28 @@
 import * as Dialog from "@radix-ui/react-dialog";
 import React from "react";
+import { ArrowLeft } from "lucide-react";
 
-import DialogContainer from "./ui/dialog-container";
+import DialogContainer from "./ui/DialogContainer";
 import { CardButton, Graphsheet, Icons } from "./ui";
 
 export interface DeleteConfirmationDialogProps {
   open: boolean;
   onClose: () => void;
-  onDone: () => void;
+  onDelete: () => void;
+  onBack: () => void;
   button: string;
+  text: string;
   heading: string;
   disableButton?: boolean;
 }
 
-const UpdateSuccessDialog: React.FC<DeleteConfirmationDialogProps> = ({
+const DeleteConfirmationDialog: React.FC<DeleteConfirmationDialogProps> = ({
   open,
   onClose,
-  onDone,
+  onDelete,
+  onBack,
   button,
-
+  text,
   heading,
   disableButton = false,
 }) => {
@@ -50,16 +54,18 @@ const UpdateSuccessDialog: React.FC<DeleteConfirmationDialogProps> = ({
                 className="absolute w-full h-full duration-500 opacity-30 z-0"
               />
               <div className="bg-white-cloud-gradient-sm absolute w-full h-full z-10" />
-
-              <Icons.Tick className="size-9 text-grey-100" />
+              <div className="h-8 w-8 bg-error-50 rounded-lg flex items-center justify-center z-20">
+                <Icons.Trash className="size-6 text-grey-100" />
+              </div>
             </div>
-            <span className="text-3xl font-medium text-grey-10 text-center mb-16">
-              {heading}
-            </span>
+            <span>{heading}</span>
           </div>
 
           {/* Mobile Header */}
           <div className="flex py-4 items-center justify-between text-grey-10 relative w-full md:hidden">
+            <button onClick={onBack} className="mr-2">
+              <ArrowLeft className="size-6 text-grey-10" />
+            </button>
             <div className="text-lg font-medium relative">
               <span className="capitalize">{heading}</span>
             </div>
@@ -68,16 +74,28 @@ const UpdateSuccessDialog: React.FC<DeleteConfirmationDialogProps> = ({
             </button>
           </div>
 
+          {/* Message */}
+          <div className="font-medium text-base text-grey-20 text-center mb-4 ">
+            {text}
+          </div>
+
           {/* Buttons */}
           <CardButton
-            className="bg-primary-50 py-4 mb-5 hover:bg-primary-40 transition text-white w-full font-medium"
+            className="bg-primary-50 py-4 hover:bg-primary-40 transition text-white w-full font-medium"
             size={"lg"}
-            variant={"dialog"}
-            onClick={onDone}
+            variant={"primary"}
+            onClick={onDelete}
             appendToStart
             disabled={disableButton}
           >
             {button}
+          </CardButton>
+          <CardButton
+            variant="secondary"
+            className="bg-grey-100 border border-grey-80 text-grey-10 w-full my-4 text-lg font-medium h-12 hover:bg-grey-80 transition"
+            onClick={onBack}
+          >
+            Go Back
           </CardButton>
         </div>
       </DialogContainer>
@@ -85,4 +103,4 @@ const UpdateSuccessDialog: React.FC<DeleteConfirmationDialogProps> = ({
   );
 };
 
-export default UpdateSuccessDialog;
+export default DeleteConfirmationDialog;
