+++ app/components/dashboard-title-wrapper/notifications-menu/NotificationOptionSelect.tsx
@@ -1,5 +1,5 @@
 import * as Menubar from "@radix-ui/react-menubar";
-import ActiveTabBg from "@/app/components/ui/tabs/active-tab-bg";
+import ActiveTabBg from "@/app/components/ui/tabs/ActiveTabBg";
 import { Icons } from "../../ui";
 
 interface Option {
@@ -46,11 +46,10 @@ const NotificationOptionSelect: React.FC<NotificationOptionSelectProps> = ({
             {options.map((opt) => (
               <Menubar.Item
                 key={opt.value}
-                className={`pl-6 pr-3 py-2 cursor-pointer text-xs leading-[18px] font-medium hover:text-primary-50 ${
-                  value === opt.value
+                className={`pl-6 pr-3 py-2 cursor-pointer text-xs leading-[18px] font-medium hover:text-primary-50 ${value === opt.value
                     ? "text-primary-50 border-l border-primary-50"
                     : "text-grey-70"
-                }`}
+                  }`}
                 onSelect={() => onChange(opt.value)}
               >
                 {opt.label}
