+++ app/components/page-sections/files/ipfs/files-table/index.tsx
@@ -29,7 +29,7 @@ import { toast } from "sonner";
 import { cn } from "@/lib/utils";
 import { usePagination } from "@/lib/hooks";
 import NameCell from "./NameCell";
-import TableActionMenu from "@/components/ui/alt-table/table-action-menu";
+import TableActionMenu from "@/app/components/ui/alt-table/TableActionMenu";
 import { getFileTypeFromExtension } from "@/lib/utils/getTileTypeFromExtension";
 import { VideoDialogTrigger } from "./VideoDialog";
 import { ImageDialogTrigger } from "./ImageDialog";
@@ -38,9 +38,9 @@ import BlockTimestamp from "@/app/components/ui/block-timestamp";
 import { Icons } from "@/app/components/ui";
 import { openUrl } from "@tauri-apps/plugin-opener";
 import { useWalletAuth } from "@/app/lib/wallet-auth-context";
-import { FileViewSharedState } from "../shared/file-view-utils";
+import { FileViewSharedState } from "../shared/FileViewUtils";
 import FileDetailsDialogContent from "../file-details-dialog-content";
-import SidebarDialog from "@/app/components/ui/sidebar-dialog";
+import SidebarDialog from "@/app/components/ui/SidebarDialog";
 
 const TIME_BEFORE_ERR = 30 * 60 * 1000;
 const columnHelper = createColumnHelper<FormattedUserIpfsFile>();
