+++ app/components/sidebar/FooterNavItems.tsx
