+++ app/components/sidebar/index.tsx
@@ -4,8 +4,8 @@ import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { AppVersion, Icons, RevealTextLine } from "../ui";
 import cn from "@/app/lib/utils/cn";
-import NavItem from "./nav-item";
-import { navItems, footerNavItems } from "./nav-data";
+import NavItem from "./NavItem";
+import { navItems, footerNavItems } from "./NavData";
 import { useAtom, useSetAtom } from "jotai";
 import {
   settingsDialogOpenAtom,
@@ -13,7 +13,7 @@ import {
   activeSettingsTabAtom
 } from "@/app/components/sidebar/sideBarAtoms";
 import { InView } from "react-intersection-observer";
-import FooterNavItem from "./footer-nav-items";
+import FooterNavItem from "./FooterNavItems";
 import SettingsWidthDialog from "../page-sections/settings/SettingsDialog";
 import SettingsDialogContent from "../page-sections/settings/SettingsDialogContent";
