+++ app/components/splash-screen/AnimatedIcons.tsx
