+++ app/components/splash-screen/AnimatedRings.tsx
