+++ app/components/splash-screen/ProgressDisplay.tsx
