+++ app/components/splash-screen/SplashContent.tsx
