+++ app/components/ui/ErrorDetail.tsx
