+++ app/components/ui/SidebarDialog.tsx
