+++ app/components/ui/alt-table/CopyableCell.tsx
