+++ app/components/ui/alt-table/TruncatedCell.tsx
