+++ app/components/ui/alt-table/index.tsx
@@ -1,10 +1,10 @@
-export * from "./td";
-export * from "./th";
-export * from "./thead";
-export * from "./tbody";
-export * from "./tablewrapper";
-export * from "./tr";
-export * from "./table";
-export * from "./copyable-cell";
-export * from "./pagination";
-export * from "./truncated-cell";
+export * from "./Td";
+export * from "./Th";
+export * from "./THead";
+export * from "./TBody";
+export * from "./TableWrapper";
+export * from "./Tr";
+export * from "./Table";
+export * from "./CopyableCell";
+export * from "./Pagination";
+export * from "./TruncatedCell";
