+++ app/components/ui/index.ts
@@ -26,12 +26,12 @@ export { default as ThreeDotLoader } from "./three-dot-loader";
 export { default as StatsBox } from "./stats-box";
 export { default as ResourceDetailsError } from "./resource-details-error";
 export { default as FileTypeTag } from "./file-type-tag";
-export { default as ErrorDetail } from "./error-detail";
+export { default as ErrorDetail } from "./ErrorDetail";
 export { default as SearchInput } from "./search-input";
 export { default as CardButton } from "./button/CardButton";
 export { default as AreaLineChart } from "./area-line-chart";
-export { default as ConfirmModal } from "./confirm-modal";
-export { default as ImportantWarnings } from "./important-warnings";
+export { default as ConfirmModal } from "./ConfirmModal";
+export { default as ImportantWarnings } from "./ImportantWarnings";
 export { default as RefreshButton } from "./refresh-button";
 export { default as InfoTooltip } from "./info-tooltip";
 export { default as FileTypeIcon } from "./file-type-icon";
@@ -39,5 +39,5 @@ export { default as WaitAMoment } from "./wait-a-moment";
 export { default as BlockTimestamp } from "./block-timestamp";
 export { default as IconButton } from "./IconButton";
 export { default as BackButton } from "./back-button";
-export { default as AppVersion } from "./app-version";
+export { default as AppVersion } from "./AppVersion";
 export { default as ChartGridOverlay } from "./chart-grid-overly";
